requirements.txt
fastapi>=0.78.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.17.6
SQLAlchemy>=1.4.36
python-dotenv>=0.20.0
psycopg2>=2.9.3